Fix execution on Windows
Firstly the program assumed the file descriptor for stdin on unix like
operating systems (which is 0). On Windows those are created per
cmd/powershell/terminal. Without this change it would crash because of
an invalid handle.
Secondly not only the stdin is needed for the call to NewTerminal() but
also stdout. It worked on unix like systems that way, but not on Windows. It
would stay with an empty prompt. With both given, the program works also
on Windows.

Fixes https://github.com/alfredxing/calc/issues/7
Fixes https://github.com/alfredxing/calc/issues/18