Shell: make use of containerd/console
It's a library which makes the heavy lifting in recognizing features of
the current active console.
term alone worked great on unix/linux based operating systems, but not
on Windows. It worked in general, but no support for arrow keys, Pos1, End and
the history. This is remedied with the use of containerd/console.

Additionally make use of the exit() function when looking for "exit/quit".
It wasn't right to simply break out of the loop as it could have had severe
side-effects. In fact with containerd/console it would noticeably break (output
send to the wrong terminal file-descriptor) and only a reset would help.
Funnily using exit() in that place will come with an additional newline. Printing
some text covers this up.