handleKey: disable it temporarily
handleKey was and is without function. ^C was and is covered by io.EOF.
It never quit at the function call.

As far as I understand every keypress is sent to the handler if used.
If so modifier keys needs to be saved in a way to construct the needed Escape sequence?

Think about if it is really needed (maybe, because DEL isn't working)
If not remove the function entirely.